Add files via upload
Assume that you are sharing a flat with your friend or some other person and you share the various bills like cooking, electricity, Bike Maintenance(maybe), Rent, etc. Each time either of you will pay the amount. You are keeping note of the expenses and at the end of the period of sharing a flat, the person who spent an excess amount shall pay the other. All this process is done manually as on date.

Being a techie, you have observed this problem and thought of making the process digital to an extent in the first version of the application. Your main goals to accomplish are,

1.	To create a json-server which can store the data for future use.
2.	Make sure every input has an id and the user should not manually insert the id for the data.
3.	To create a GUI structure where the data from json-server can be seen.
4.	To create a GUI structure where you can fill the details needed to store in the server.
5.	At the end of the data retrieved from server, show Total spent, person A spent, Person B spent, who pays to whom and how has to be paid by them at any given time.
6.	View the Add new Item Form separately with the react router dom.

 

 

 


Future Developments: 
1.	Create a functionality where you can select a particular month data
2.	Restrict the usage of the date of the previous month.
3.	Create Login functionality. 
4.	Restrict server update of previous month.